requirements.txt
torch>=2.2
torchvision
hydra-core>=1.3
omegaconf
numpy
pandas
scipy
tqdm
h5py
